+++ mobile/lib/screens/record_list/record_list.dart
@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
 import 'package:mobile/models/api_models/record_model.dart';
 import 'package:mobile/models/dialog_results/record_key_dialog_result.dart';
-import 'package:mobile/models/hive_models/stored_record.dart';
 import 'package:mobile/screens/record_add/record_add.dart';
 import 'package:mobile/screens/record_list/record_key_dialog/record_key_dialog.dart';
 import 'package:mobile/screens/record_list/record_tile.dart';
+import 'package:mobile/services/record_offline_service.dart';
 import 'package:mobile/services/record_service.dart';
 import 'package:mobile/models/service_results/result.dart';
 import 'package:mobile/utils/box_names.dart';
@@ -24,17 +24,15 @@ class RecordList extends StatefulWidget {
 class _RecordListState extends State<RecordList> {
   late Future<Result<List<RecordModel>>> futureRecords = Future.value(Result([]));
   late Queue<Exception>? errors = Queue();
+  RecordOfflineService recordOfflineService = RecordOfflineService();
   String guid = "";
-  bool offlineMode = false;
+  bool offlineMode = true;
 
   @override
   void initState() {
     super.initState();
     fetchRecords();
   }
-
-  String getOfflineBoxKey(String guid) => "${guid}_offline";
-  String getOnlineBoxKey(String guid) => "${guid}_online";
 
   Future<void> fetchRecords() async {
     Result<bool> isValidGuid = await validateGuid(guid);
@@ -70,73 +68,23 @@ class _RecordListState extends State<RecordList> {
   }
 
   Future<Result<List<RecordModel>>> fetchCombinedRecordsOffline() async {
-    var lastOnlineRecords = await fetchLastOnlineRecords();
-    return combineAndSortRecords(await fetchOfflineRecords(), lastOnlineRecords.value);
+    var lastOnlineRecords = await recordOfflineService.fetchLastOnlineRecords(guid);
+    return combineAndSortRecords(await recordOfflineService.fetchOfflineRecords(guid), lastOnlineRecords.value);
   }
 
   Future<Result<List<RecordModel>>> fetchCombinedRecordsOnline() async {
     var onlineRecords = await RecordService().fetchRecords(guid);
-    if (await isGuidStored(guid)) {
-      storeLastOnlineData(onlineRecords);
+    if (await recordOfflineService.isGuidStored(guid)) {
+      recordOfflineService.storeLastOnlineData(guid, onlineRecords);
     }
 
-    return combineAndSortRecords(await fetchOfflineRecords(), onlineRecords.value);
+    return combineAndSortRecords(await recordOfflineService.fetchOfflineRecords(guid), onlineRecords.value);
   }
 
   Future<Result<List<RecordModel>>> combineAndSortRecords(List<RecordModel> recordA, List<RecordModel> recordB) {
     var result = Result([...recordA, ...recordB]);
     result.value.sort((a, b) => a.createdAt.compareTo(b.createdAt));
     return Future.value(result);
-  }
-
-  Future<void> storeLastOnlineData(Result<List<RecordModel>> result) async {
-    final box = await Hive.openBox(getOnlineBoxKey(guid));
-    await box.clear();
-    await box.addAll(result.value.map((record) => StoredRecord(
-        guid: guid,
-        reason: record.reason,
-        amount: record.amount,
-        receipt: record.receipt,
-        signature: record.signature,
-        createdAt: record.createdAt,
-        isSynced: true,
-      )));
-  }
-
-  Future<Result<List<RecordModel>>> fetchLastOnlineRecords() async {
-    // Get offline records if any
-    List<RecordModel> onlineRecords = [];
-    if (await isGuidStored(guid)) {
-      final onlineKey = getOnlineBoxKey(guid);
-      final boxExists = await Hive.boxExists(onlineKey);
-      if (boxExists) {
-        final box = await Hive.openBox(onlineKey);
-        onlineRecords = box.values
-            .map((val) => RecordModel.fromLocal(val, true))
-            .toList();
-      }
-    }
-    var result = Result(onlineRecords);
-    return result;
-  }
-
-  Future<List<RecordModel>> fetchOfflineRecords() async {
-    List<RecordModel> offlineRecords = [];
-    if (await isGuidStored(guid)) {
-      final boxExists = await Hive.boxExists(getOfflineBoxKey(guid));
-      if (boxExists) {
-        final box = await Hive.openBox(getOfflineBoxKey(guid));
-        offlineRecords = box.values
-            .map((val) => RecordModel.fromLocal(val, false))
-            .toList();
-      }
-    }
-    return offlineRecords;
-  }
-
-  Future<bool> isGuidStored(String guid) async {
-    var keyBox = await Hive.openBox(BoxNames.keys);
-    return keyBox.values.any((key) => key == guid);
   }
 
   void retrieveGuid() async {
@@ -149,7 +97,7 @@ class _RecordListState extends State<RecordList> {
     
     if(guid.isNotEmpty) {
       if(currentName != null) {
-        var keyBox = Hive.box(BoxNames.keys);
+        var keyBox = await Hive.openBox(BoxNames.keys);
         keyBox.put(currentName, guid);
       }
